requirements.txt
Django==2.0.1
django-cors-headers==2.1.0
djangorestframework==3.7.7
djangorestframework-camel-case==0.2.0
psycopg2==2.7.3.2
pytz==2017.3